requirements.txt
Flask
pandas
openpyxl
Werkzeug
python-dotenv
flask
apscheduler
python-telegram-bot==13.15
